refactor: move util func in sub packages
