+++ pkg/engine/jsonutils/traverse.go
@@ -5,7 +5,7 @@ import (
 	"strconv"
 	"strings"
 
-	"github.com/kyverno/kyverno/pkg/utils"
+	datautils "github.com/kyverno/kyverno/pkg/utils/data"
 	"golang.org/x/exp/slices"
 )
 
@@ -71,13 +71,13 @@ func (t *Traversal) traverseJSON(element interface{}, path string) (interface{},
 	// traverse further
 	switch typed := element.(type) {
 	case map[string]interface{}:
-		return t.traverseObject(utils.CopyMap(typed), path)
+		return t.traverseObject(datautils.CopyMap(typed), path)
 
 	case []interface{}:
 		return t.traverseList(slices.Clone(typed), path)
 
 	case []map[string]interface{}:
-		return t.traverseList(utils.CopySliceOfMaps(typed), path)
+		return t.traverseList(datautils.CopySliceOfMaps(typed), path)
 
 	case Key:
 		return typed.Key, nil
